String formatting functions isolated to seperate file. This also includes a basic fix for under/over flow testing.
